Simplifying some code
