+++ server/src-lib/Hasura/GraphQL/Transport/HTTP.hs
@@ -50,7 +50,7 @@ class Monad m => MonadExecuteQuery m where
   executeQuery
     :: GQLReqParsed
     -> [QueryRootField UnpreparedValue]
-    -> Maybe EQ.PreparedSql -- EQ.GeneratedSqlMap -- HashMap G.Name EQ.PreparedSql
+    -> Maybe EQ.PreparedSql
     -> PGExecCtx
     -> Q.TxAccess
     -> TraceT (LazyTx QErr) EncJSON
@@ -261,43 +261,3 @@ runMutationDB reqId query userInfo tx =  do
   !resp <- liftEither respE
   let telemQueryType = Telem.Mutation
   return (telemTimeIO, telemQueryType, resp)
-
-{-
-runHasuraGQ
-  :: ( MonadIO m
-     , MonadError QErr m
-     , MonadReader E.ExecutionCtx m
-     , MonadQueryLog m
-     , MonadTrace m
-     , MonadExecuteQuery m
-     )
-  => RequestId
-  -> (GQLReqUnparsed, GQLReqParsed)
-  -> UserInfo
-  -> E.ExecOp (Tracing.TraceT (LazyTx QErr))
-  -> m (DiffTime, Telem.QueryType, HTTP.ResponseHeaders, EncJSON)
-  -- ^ Also return 'Mutation' when the operation was a mutation, and the time
-  -- spent in the PG query; for telemetry.
-runHasuraGQ reqId (query, queryParsed) userInfo resolvedOp = do
-  (E.ExecutionCtx logger _ pgExecCtx _ _ _ _ _) <- ask
-  (telemTimeIO, respE) <- withElapsedTime $ runExceptT $ case resolvedOp of
-    E.ExOpQuery tx genSql asts -> trace "Query" $ do
-      -- log the generated SQL and the graphql query
-      logQueryLog logger query genSql reqId
-      Tracing.interpTraceT id $ executeQuery queryParsed asts genSql pgExecCtx Q.ReadOnly tx
-
-    E.ExOpMutation respHeaders tx -> trace "Mutate" $ do
-      logQueryLog logger query Nothing reqId
-      ctx <- Tracing.currentContext
-      (respHeaders,) <$>
-        Tracing.interpTraceT (runLazyTx pgExecCtx Q.ReadWrite . withTraceContext ctx . withUserInfo userInfo) tx
-
-    E.ExOpSubs _ ->
-      throw400 UnexpectedPayload
-      "subscriptions are not supported over HTTP, use websockets instead"
-
-  (respHdrs, resp) <- liftEither respE
-  let !json = encodeGQResp $ GQSuccess $ encJToLBS resp
-      telemQueryType = case resolvedOp of E.ExOpMutation{} -> Telem.Mutation ; _ -> Telem.Query
-  return (telemTimeIO, telemQueryType, respHdrs, json)
--}
